requirements.txt
neat-python
pygame~=2.5.0

pip~=22.3.1
wheel~=0.38.4
numpy~=1.25.1
setuptools~=65.5.1